Add functionality to LFO

+++ Code/Plugin/Source/Wavetable.h
@@ -12,17 +12,17 @@ class SinWavetable : public IWavetable
 public:
 	SinWavetable(int length);
 	virtual ~SinWavetable();
-	virtual double getSample(int idx, double frequency) const
+	virtual double getSample(double idx, double frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
-	virtual float getSample(int idx, float frequency) const
+	virtual float getSample(double idx, float frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
 private:
 	template<typename T>
-	T __getSample(int idx, T frequency) const;
+	T __getSample(double idx, T frequency) const;
 	double __table[size];
 };
 
@@ -31,17 +31,17 @@ class SquareWavetable : public IWavetable
 public:
 	SquareWavetable(int size, double sampleRate);
 	virtual ~SquareWavetable();
-	virtual double getSample(int idx, double frequency) const
+	virtual double getSample(double idx, double frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
-	virtual float getSample(int idx, float frequency) const
+	virtual float getSample(double idx, float frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
 private:
 	template<typename T>
-	T __getSample(int idx, T frequency) const;
+	T __getSample(double idx, T frequency) const;
 	const int __nrOfTables = 10;
 	double __tables[10][size];
 };
@@ -51,17 +51,17 @@ class SawWavetable : public IWavetable
 public:
 	SawWavetable(int size, double sampleRate);
 	virtual ~SawWavetable();
-	virtual double getSample(int idx, double frequency) const
+	virtual double getSample(double idx, double frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
-	virtual float getSample(int idx, float frequency) const
+	virtual float getSample(double idx, float frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
 private:
 	template<typename T>
-	T __getSample(int idx, T frequency) const;
+	T __getSample(double idx, T frequency) const;
 	const int __nrOfTables = 10;
 	double __tables[10][size];
 };
@@ -71,17 +71,17 @@ class TriangleWavetable : public IWavetable
 public:
 	TriangleWavetable(int size, double sampleRate);
 	virtual ~TriangleWavetable();
-	virtual double getSample(int idx, double frequency) const
+	virtual double getSample(double idx, double frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
-	virtual float getSample(int idx, float frequency) const
+	virtual float getSample(double idx, float frequency) const
 	{
 		return __getSample(idx, frequency);
 	}
 private:
 	template<typename T>
-	T __getSample(int idx, T frequency) const;
+	T __getSample(double idx, T frequency) const;
 	const int __nrOfTables = 10;
 	double __tables[10][size];
 };
